+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Editor
   - Adding require command fails for requires without alias. #1791
   - Add require command without alias now add requires with brackets.
+  - Add code action "Add require" for namespaced keywords. #1793
 
 ## 2024.03.13-13.11.00