+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2024.03.31-19.10.13
+
 - Editor
   - Adding require command fails for requires without alias. #1791
   - Add require command without alias now add requires with brackets.